requirements.txt
aiogram~=2.21
environs~=9.5.0
translators~=5.4.2
Pillow~=9.2.0